requirements.txt
aliyun-python-sdk-core==2.13.4
aliyunsdkcore==1.0.3
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.1
django-cors-headers==3.0.0
django-filter==2.1.0
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
drf-writable-nested==0.5.1
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
Markdown==3.1
MarkupSafe==1.1.1
pyasn1==0.4.5
pycrypto==2.6.1
pycryptodomex==3.7.2
PyJWT==1.7.1
python-alipay-sdk==1.10.0
pytz==2019.1
requests==2.21.0
rsa==4.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.3